LayoutTests/fast/forms/ios/form-control-refresh/select/all-options-in-single-optgroup-expected.txt

This test verifies that selecting an option when all options are in a single optgroup works as expected.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS select.value is "January"
PASS select.value is "April"
PASS select.value is "October"
PASS successfullyParsed is true

TEST COMPLETE
